Configurable ARX rounds, fix error printing, simplify protobufs

+++ src/utils/threading.rs
@@ -66,11 +66,12 @@ impl Semaphore {
 }
 
 pub fn get_worker_slice<T>(max_value: T, worker_id: u32, worker_total: u32) -> (T, T) where
-    T: TryFrom<usize> + Into<usize> + Copy,
-    <T as TryFrom<usize>>::Error: std::fmt::Debug
+    T: TryFrom<u64> + Into<u64> + Copy,
+    <T as TryFrom<u64>>::Error: std::fmt::Debug
 {
-    let min = (worker_id as usize * (T::into(max_value) + 1)) / worker_total as usize;
-    let max = ((worker_id + 1) as usize * (T::into(max_value) + 1)) / worker_total as usize;
+    // TODO this is a really shitty method. improve it
+    let min = (worker_id as u64 * (T::into(max_value) + 1)) / worker_total as u64;
+    let max = ((worker_id + 1) as u64 * (T::into(max_value) + 1)) / worker_total as u64;
     (T::try_from(min).unwrap(), T::try_from(max - 1).unwrap())
 }
